[](refact) Update #3
