+++ render/src/lib/views/sidebar.vertical/port.available/styles.less
@@ -37,18 +37,6 @@
     & p {
         margin-right: 20px;
     }
-    & p.indicator {
-        margin-top: 13px;
-        height: 12px;
-        width: 12px;
-        border-radius: 50%;
-    }
-    & p.connected {
-        background-color: @scheme-color-accent;
-    }
-    & p.disconnected {
-        background-color: @scheme-color-2;
-    }
     & mat-icon {
         padding-top: 7px;
     }
