--- render/src/src/lib/common/interface.portinfo.ts
@@ -1,24 +1,0 @@
-export interface IPortInfo {
-    path: string;
-    manufacturer?: string;
-    serialNumber?: string;
-    pnpId?: string;
-    locationId?: string;
-    productId?: string;
-    vendorId?: string;
-}
-
-export interface IIOState {
-    read: number;
-    written: number;
-}
-
-export interface IPortState {
-    ioState: IIOState;
-    connections: number;
-}
-
-export interface IPortSession {
-    default: string;
-    ports: IPortInfo[];
-}
