--- render/src/src/lib/views/row/render.api.ts
@@ -1,22 +1,0 @@
-// tslint:disable:no-inferrable-types
-
-import * as Toolkit from 'chipmunk.client.toolkit';
-
-export class SerialPortRowRenderAPI extends Toolkit.TypedRowRenderAPIExternal {
-
-    private _selector: string = 'lib-serial-row-component';
-
-    constructor() {
-        super();
-    }
-
-    public getSelector(): string {
-        return this._selector;
-    }
-    public getInputs(): { [key: string]: any } {
-        return {
-            service: null,
-        };
-    }
-
-}
