--- render/src/src/lib/views/row/styles.less
@@ -1,35 +1,0 @@
-
-@import '../../../../../../theme/variables.less';
-
-:host {
-    position: relative;
-    display: inline-block;
-    padding: 0;
-    margin: 0;
-    white-space: nowrap;
-    & span.icon{
-        position: relative;
-        display: inline-block;
-        padding: 0 4px 0 0;
-        z-index: 1;
-        cursor: default;
-    }
-    & span.content{
-        position: relative;
-        display: inline-block;
-        font-family: 'console', monospace;
-        font-size: 13.5px;
-        font-variant-numeric: tabular-nums;
-        padding: 0;
-        margin: 0;
-        line-height: 16px;
-        height: 16px;
-        max-height: 16px;
-        white-space: pre;
-        vertical-align: top;
-        z-index: 0;
-        width: 100%;
-        user-select: text;
-        font-weight: 500;
-    }
-}
