--- render/src/src/lib/views/row/template.html
@@ -1,2 +1,0 @@
-<span *ngIf="_ng_title !== ''" class="icon fas fa-plug" [attr.title]="_ng_title" [ngStyle]="{'color': _ng_color}"></span>
-<span class="content" [innerHTML]="_ng_html"></span>
