--- render/src/src/lib/views/sidebar.vertical/port.available/styles.less
@@ -1,48 +1,0 @@
-@import '../../../../../../../theme/variables.less';
-
-:host {
-    & div.container {
-        position: relative;
-        overflow: hidden;
-    }
-    & div.info {
-        justify-content: space-between;
-        display: flex;
-    }
-    & div.canvas {
-        text-align:left;
-        position: relative;
-        width: 100%;
-        height: 25px;
-        margin-bottom: 16px;
-        & div.holder {
-            position: absolute;
-            top: 0;
-            right: 0;
-            bottom: 0;
-            left: -30px;
-        }
-        &:after {
-            position: absolute;
-            content: '';
-            width: 100%;
-            height: 1px;
-            bottom: -1px;
-            border-bottom: 1px solid @scheme-color-2;
-        }
-    }
-    & div.right {
-        display: flex;
-    }
-    & p {
-        margin-right: 20px;
-    }
-    & mat-icon {
-        padding-top: 7px;
-    }
-    & input {
-        width: calc(100% - 8px);
-        background: transparent;
-        padding-top: 7px;
-    }
-}
