--- render/src/src/lib/views/sidebar.vertical/port.connected/styles.less
@@ -1,36 +1,0 @@
-
-:host {
-    position: relative;
-    display: block;
-    width: 100%;
-    & div.toggle {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 1rem;
-        height: 100%;
-        z-index: 2;
-        & span.toggle-button{
-            position: absolute;
-            top: 1px;
-            left: -2px;
-            transition: all 100ms ease-in-out;
-            &.opened{
-                transform: rotateZ(180deg);
-            }
-        }
-    }
-    & div.info{
-        position: relative;
-        padding: 0 0 0 1rem;
-        & span.size{
-            position: absolute;
-            display: block;
-            right: 0;
-            top: 0;
-        }
-        & div.controlls {
-            text-align: right;
-        }
-    }
-}
