--- render/src/src/lib/views/sidebar.vertical/port.connected/template.html
@@ -1,15 +1,0 @@
-<div class="toggle">
-    <span   [attr.class]="'toggle-button small-icon-button fas fa-angle-double-down' + (_ng_isMoreOpened() ? ' opened' : '')"
-            [attr.title]="'Show/Hide more information'"
-            (click)="_ng_onMore($event)"></span>
-</div>
-<div class="info">
-    <p class="t-small color-scheme-1">{{port.path}}</p>
-    <span class="size t-small color-scheme-2">{{_ng_read}}</span>
-    <div class="more" *ngIf="_ng_isMoreOpened()">
-        <lib-sb-port-options-read-com *ngIf="options !== undefined" [options]="options"></lib-sb-port-options-read-com>
-        <div class="controlls">
-            <span class="small-button" (click)="_ng_onDisconnect()">Disconnect</span>
-        </div>
-    </div>
-</div>
