--- render/src/src/lib/views/sidebar.vertical/port.listed/template.html
@@ -1,7 +1,0 @@
-<div [attr.class]="'info ' + (_ng_more.length > 0 ? 'hasinfo' : '')">
-    <p class="t-small color-scheme-1">{{port.path}}</p>
-    <span class="size t-small color-scheme-2">{{_ng_read}}</span>
-    <div class="more" *ngIf="_ng_isMoreOpened()">
-        <p class="t-small color-scheme-2" *ngFor="let item of _ng_more">{{item.name}}: {{item.value}}</p>
-    </div>
-</div>
