--- render/src/src/lib/views/sidebar.vertical/port.options.read/template.html
@@ -1,6 +1,0 @@
-<table>
-    <tr *ngFor="let option of _ng_options">
-        <td class="t-small color-scheme-2">{{option.key}}</td>
-        <td class="t-small color-scheme-2">{{option.value}}</td>
-    </tr>
-</table>
