--- render/src/src/lib/views/sidebar.vertical/port.options.write/styles.less
@@ -1,32 +1,0 @@
-
-:host {
-    position: relative;
-    display: block;
-    width: 100%;
-    & ul.settings{
-        position: relative;
-        display: block;
-        padding: 0;
-        margin: 0;
-        list-style: none;
-        & li.setting{
-            position: relative;
-            display: block;
-            padding: 0;
-            margin: 0;
-            list-style: none;
-        }
-    }
-    & span {
-        margin-left: 2%;
-    }
-    & span.delimiter{
-        position: relative;
-        display: block;
-        height: 0.5rem;
-        width: 1px;
-    }
-    & mat-error {
-        color: #ff0000;
-    }
-}
