--- render/src/src/lib/views/sidebar.vertical/port.options.write/template.html
@@ -1,89 +1,0 @@
-<mat-form-field>
-    <mat-label>Baud rate</mat-label>
-    <mat-select [(ngModel)]="baudRate" [placeholder]="baudRatePlaceholder" (selectionChange)="_ng_change(); _ng_onBRChange()">
-        <mat-option *ngFor="let baudRateItem of _ng_baudrateItems" [value]="baudRateItem.value">
-            {{baudRateItem.caption}}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
-
-<mat-form-field *ngIf="custom">
-    <input matInput type="'text'" [formControl]="baudRateInput" [placeholder]="'Type your custom baud rate'" [value]="(baudRate === -1) ? baudRateInput.value : baudRate" (keyup)="_ng_change()"/>
-    <mat-error *ngIf="baudRateInput.hasError('pattern')">
-        Please enter a valid Baud rate
-    </mat-error>
-    <mat-error *ngIf="baudRateInput.hasError('required')">
-        Baud rate is required!
-    </mat-error>
-</mat-form-field>
-
-<mat-form-field>
-    <mat-label>Data bits</mat-label>
-    <mat-select [(ngModel)]="dataBits" [placeholder]="dataBits" (selectionChange)="_ng_change()">
-        <mat-option *ngFor="let databitsItem of _ng_databitsItems" [value]="databitsItem.value">
-            {{databitsItem.caption}}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
-
-<mat-form-field>
-    <mat-label>Stop bits</mat-label>
-    <mat-select [(ngModel)]="stopBits" [placeholder]="stopBits" (selectionChange)="_ng_change()">
-        <mat-option *ngFor="let stopbitsItem of _ng_stopbitsItems" [value]="stopbitsItem.value">
-            {{stopbitsItem.caption}}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
-
-<mat-form-field>
-    <mat-label>Parity</mat-label>
-    <mat-select [(ngModel)]="parity" [placeholder]="parity" (selectionChange)="_ng_change()">
-        <mat-option *ngFor="let parityItem of _ng_parityItems" [value]="parityItem.value">
-            {{parityItem.caption}}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
-
-<mat-form-field>
-    <mat-label>Encoding</mat-label>
-    <mat-select [(ngModel)]="encoding" [placeholder]="encoding" (selectionChange)="_ng_change()">
-        <mat-option *ngFor="let encodingItem of _ng_encodingItems" [value]="encodingItem.value">
-            {{encodingItem.caption}}
-        </mat-option>
-    </mat-select>
-</mat-form-field>
-
-<mat-form-field>
-    <input matInput [(ngModel)]="highWaterMark" type="'number'" (keyup)="_ng_change()" [placeholder]="'High water mark, bytes'"/>
-</mat-form-field>
-
-<mat-form-field>
-    <input matInput [(ngModel)]="delimiter" type="'text'" (keyup)="_ng_change()" [placeholder]="'Line delimiter'"/>
-</mat-form-field>
-
-<mat-list>
-    <mat-list-item>
-        <mat-checkbox [(ngModel)]="lock" (change)="_ng_change()"></mat-checkbox>
-        <span>Lock port</span>
-    </mat-list-item>
-
-    <mat-list-item>
-        <mat-checkbox [(ngModel)]="rtscts" (change)="_ng_change()"></mat-checkbox>
-        <span>RTSCTS</span>
-    </mat-list-item>
-
-    <mat-list-item>
-        <mat-checkbox [(ngModel)]="xon" [checked]="false" (change)="_ng_change()"></mat-checkbox>
-        <span>xON</span>
-    </mat-list-item>
-
-    <mat-list-item>
-        <mat-checkbox [(ngModel)]="xoff" [checked]="false" (change)="_ng_change()"></mat-checkbox>
-        <span>xOFF</span>
-    </mat-list-item>
-
-    <mat-list-item>
-        <mat-checkbox [(ngModel)]="xany" [checked]="false" (change)="_ng_change()"></mat-checkbox>
-        <span>xANY</span>
-    </mat-list-item>
-</mat-list>
