--- render/src/src/lib/views/sidebar.vertical/port.options/styles.less
@@ -1,3 +1,0 @@
-& button {
-    margin: 3%;
-}
