--- render/src/src/lib/views/sidebar.vertical/styles.less
@@ -1,45 +1,0 @@
-@import '../../../../../../theme/variables.less';
-
-:host {
-    position: absolute;
-    display: block;
-    top:0.5rem;
-    left:0.5rem;
-    right: 0.5rem;
-    bottom: 0.5rem;
-    overflow-x: hidden;
-    overflow-y: auto;
-    & ul.ports {
-        position: relative;
-        display: block;
-        width: 100%;
-        list-style: none;
-        margin: 0;
-        padding: 0;
-        & li.port{
-            position: relative;
-            display: block;
-            width: 100%;
-            list-style: none;
-            margin: 0;
-            padding: 3px 0;
-            &.selected{
-                background: @scheme-color-4;
-            }
-            &:hover{
-                background: @scheme-color-4;
-            }
-        }
-        & button {
-            margin-top: 10px;
-            width: 100%;
-        }
-        &.connected{
-            & li.port{
-                &:hover{
-                    background: none;
-                }
-            }
-        }
-    }
-}
