--- render/src/src/lib/views/sidebar.vertical/template.html
@@ -1,18 +1,0 @@
-<mat-expansion-panel
-    [disabled]="true"
-    [hideToggle]="true"
-    [expanded]="true">
-    <mat-expansion-panel-header>
-        <mat-panel-title>
-            Available ports
-        </mat-panel-title>
-    </mat-expansion-panel-header>
-    <ul class="ports" *ngIf="_ng_ports.length > 0">
-        <li *ngFor="let port of _ng_ports">
-            <button mat-raised-button>
-                <lib-dia-port-available-com [port]="port" [observables]="subjects">
-                </lib-dia-port-available-com>
-            </button>
-        </li>
-    </ul>    
-</mat-expansion-panel>
