--- render/src/src/public_api.ts
@@ -1,16 +1,0 @@
-/*
- * Public API Surface of terminal
- */
-/*
- * Public API Surface of terminal
- */
-import { SerialPortRowRender } from './lib/views/row/render';
-import Service from './lib/services/service';
-
-const externalRowRender = new SerialPortRowRender();
-
-export { externalRowRender };
-
-export * from './lib/views/sidebar.vertical/component';
-export * from './lib/module';
-export { Service };
